cli(progress): split ci-mode state into own module
progress.rs was 974 lines mixing the TTY dispatcher, CI heartbeat, and
the `PausingWriter` tracing sink. Move the CI-mode state + rendering
helpers into `progress/ci.rs` so the two UI modes live in separate
files and the shared dispatcher reads top-to-bottom. No behavior change.
